requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
allennlp==2.10.1
altair==4.2.2
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
backports.tarfile==1.2.0
base58==2.1.1
bcrypt==4.2.0
blinker==1.8.2
blis==0.7.11
boto3==1.34.144
botocore==1.34.144
build==1.2.1
cached-path==1.1.6
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
Chroma==0.2.0
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
commonmark==0.9.1
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.5.14
debugpy==1.8.2
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
fairscale==0.4.6
faiss-cpu==1.8.0.post1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.7.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gast==0.4.0
gensim==3.8.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
gTTS==2.3.1
gym==0.25.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
jaraco.context==5.3.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
JPype1==1.5.0
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.11.0
kiwisolver==1.4.5
konlpy==0.6.0
kubernetes==30.1.0
langchain==0.0.181
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
llama-index==0.6.12
lmdb==1.5.1
lxml==5.2.2
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
music21==9.1.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.8.1
numexpr==2.10.1
numpy==1.24.2
oauthlib==3.2.2
onnxruntime==1.18.1
openai==0.28.0
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==1.5.3
parso==0.8.4
pathlib_abc==0.1.1
pathtools==0.1.2
pathy==0.11.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
posthog==3.5.0
preprocess==2.0.0
preshed==3.0.9
promise==2.3
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==3.20.3
psutil==6.0.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.17
pydeck==0.9.1
pygame==2.6.0
Pygments==2.18.0
Pympler==1.1
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==12.6.0
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
sentry-sdk==2.10.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
spacy==3.3.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.24
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
streamlit==1.20.0
streamlit-audiorecorder==0.0.2
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
thinc==8.0.17
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==1.12.1
torchvision==0.13.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.43.2
typer==0.12.3
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.30.3
validators==0.33.0
wandb==0.12.21
wasabi==0.10.1
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wolframalpha==5.1.3
wordcloud==1.9.3
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
zipp==3.19.2
